+++ tops_backend/src/memory/mod.rs
@@ -63,7 +63,7 @@ impl MemCpyTrait for TopsMemCpy {
         src_ptr: *const c_void,
         size: usize,
     ) -> DeviceResult<()> {
-        driv::topsMemcpyHtoD(d_ptr, src_ptr as *mut c_void, size as u64).to_result()?;
+        driv::topsMemcpyHtoD(d_ptr, src_ptr as *mut c_void, size).to_result()?;
         Ok(())
     }
 
@@ -74,7 +74,7 @@ impl MemCpyTrait for TopsMemCpy {
         src_ptr: Self::RawDevicePointerT,
         size: usize,
     ) -> DeviceResult<()> {
-        driv::topsMemcpyDtoH(d_ptr, src_ptr, size as u64).to_result()?;
+        driv::topsMemcpyDtoH(d_ptr, src_ptr, size).to_result()?;
         Ok(())
     }
 
@@ -84,7 +84,7 @@ impl MemCpyTrait for TopsMemCpy {
     /// Returns in `.0` the amount of memory on the device that is free according to
     /// the OS. Device is not guaranteed to be able to allocate all of the memory that
     /// the OS reports as free.
-    fn mem_get_info() -> DeviceResult<(u64, u64)> {
+    fn mem_get_info() -> DeviceResult<(usize, usize)> {
         let mut mem_free = 0;
         let mut mem_total = 0;
         unsafe {
